Moved generic tests to conftest.py
