+++ tests/test_sqlite3.py
@@ -1,4 +1,4 @@
-from pathlib import Path, PosixPath
+from pathlib import Path
 from typing import NamedTuple
 
 import aiosql
@@ -24,27 +24,6 @@ def dict_factory(cursor, row):
 def queries():
     p = Path(__file__).parent / "blogdb" / "sql"
     return aiosql.from_path(p, "sqlite3", RECORD_CLASSES)
-
-
-blogs_files_fn = [
-    ("blogs", "blogs.sql", ["publish_blog", "remove_blog", "get_user_blogs"]),
-    (
-        "blogs",
-        "blogs_pg.sql",
-        ["pg_get_blogs_published_after", "pg_publish_blog", "pg_bulk_publish"],
-    ),
-    ("blogs", "blogs_sqlite.sql", ["sqlite_get_blogs_published_after", "sqlite_bulk_publish"]),
-    ("users", "users.sql", ["get_all", "get_by_username", "get_by_lastname", "get_all_sorted"]),
-]
-
-
-@pytest.mark.parametrize("dir, file, functions", blogs_files_fn)
-def test_introspected_queries_path(sqlite3_conn, queries, dir, file, functions):
-    for fn in functions:
-        assert (
-            queries.__getattribute__(dir).__getattribute__(fn).file_path
-            == Path(__file__).parent / "blogdb" / "sql" / dir / file
-        )
 
 
 def test_record_query(sqlite3_conn, queries):
